requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.9.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-apscheduler==0.6.2
django-celery==3.3.1
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
importlib-metadata==4.11.4
kombu==3.0.37
Markdown==3.3.7
pipreqs==0.4.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarg==0.1.9
zipp==3.8.0
